+++ Drivers/Levoit/LevoitCore300S.groovy
@@ -722,13 +722,13 @@ def checkHttpResponse(action, resp) {
 		return true
 	else if (resp.status == 400 || resp.status == 401 || resp.status == 404 || resp.status == 409 || resp.status == 500)
 	{
-		log.error "${action}: ${resp.status} - ${resp.getData()}"
+		logError "${action}: ${resp.status} - ${resp.getData()}"
 		recordError("${action}: ${resp.status}", [site:"checkHttpResponse"])
 		return false
 	}
 	else
 	{
-		log.error "${action}: unexpected HTTP response: ${resp.status}"
+		logError "${action}: unexpected HTTP response: ${resp.status}"
 		recordError("${action}: unexpected HTTP response: ${resp.status}", [site:"checkHttpResponse"])
 		return false
 	}